Attempt at Daily Temperatures problem
The problem asks to return an array with how many days you would have to wait until
a warmer temperature for each day. My initial thinking was to loop through the given
array and count the number of days until the current day's temperature is no longer
greater than the subsequent days and break out of the loop.